EnhancedTokenomic Update

+++ TeachCrowdSale.Core/Data/Entities/BurnSnapshot.cs
@@ -9,9 +9,9 @@ namespace TeachCrowdSale.Core.Data.Entities
     public class BurnSnapshot
     {
         public int Id { get; set; }
-        public long TotalBurned { get; set; }
+        public decimal TotalBurned { get; set; }
         public decimal BurnedPercentage { get; set; }
-        public long BurnedLast30Days { get; set; }
+        public decimal BurnedLast30Days { get; set; }
         public decimal BurnRate { get; set; }
         public decimal EstimatedAnnualBurn { get; set; }
         public DateTime LastBurnDate { get; set; }
