+++ TeachCrowdSale.Core/Models/Burning/BurnStatisticsModel.cs
@@ -8,9 +8,9 @@ namespace TeachCrowdSale.Core.Models.Burning
 {
     public class BurnStatisticsModel
     {
-        public long TotalBurned { get; set; }
+        public decimal TotalBurned { get; set; }
         public decimal BurnedPercentage { get; set; }
-        public long BurnedLast30Days { get; set; }
+        public decimal BurnedLast30Days { get; set; }
         public decimal BurnRate { get; set; }
         public decimal EstimatedAnnualBurn { get; set; }
         public DateTime LastBurnDate { get; set; }
