+++ TeachCrowdSale.Core/Models/Tokenomics/SupplyAllocationModel.cs
@@ -9,7 +9,7 @@ namespace TeachCrowdSale.Core.Models.Tokenomics
     public class SupplyAllocationModel
     {
         public string Category { get; set; } = string.Empty;
-        public long TokenAmount { get; set; }
+        public decimal TokenAmount { get; set; }
         public decimal Percentage { get; set; }
         public string Color { get; set; } = string.Empty;
         public string Description { get; set; } = string.Empty;
